+++ app/routes/today.tsx
@@ -1,4 +1,4 @@
-import Reservation from '../../components/Reservation';
+import Reservation from '../components/Reservation';
 
 export default function Today() {
     const time = new Date();
@@ -6,6 +6,8 @@ export default function Today() {
     var events = [Reservation({title: "Event 1", room: "W210", start: time, end: time}, true), Reservation({title: "Event 2", room: "W212", start: time2, end: time2}, true)];
     return <>
         <h1>Welcome to Today! {time.toLocaleDateString("en-US", {month: "long", day:"numeric"})}</h1>
-        {events.map((event) => event)}
+        <ul>
+            {events.map((event) => event)}
+        </ul>
     </>
 }
